controllight_update: u8 bit local drops redundant clrlwi+mr copy (89.9->92.3%)

+++ src/main/dll/LGT/dll_02AC_lgtcontrollight.c
@@ -78,11 +78,11 @@ void controllight_init(int obj, int setup)
 void controllight_update(int obj)
 {
     extern void pointlight_setEffectState(int obj, int enabled);
-    int bit;
+    u8 bit;
     ControlLightState* state = ((GameObject*)obj)->extra;
-    bit = (u8)GameBit_Get(state->gameBit);
+    bit = GameBit_Get(state->gameBit);
 
-    if ((u32)bit != state->lastBit)
+    if (bit != state->lastBit)
     {
         switch (state->invertMode)
         {
